Refactor seeding file to automatically diconnect from the database after completion
