+++ package.json
@@ -18,6 +18,7 @@
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
     "babel-loader": "^8.1.0",
+    "body-parser": "^1.19.0",
     "express": "^4.17.1",
     "faker": "^5.1.0",
     "mongoose": "^5.10.6",
